+++ benchmarks/main.rs
@@ -17,7 +17,6 @@
 
 mod barrier;
 mod blocking;
-mod broadcast;
 mod condvar;
 mod latch;
 mod mpsc;