+++ tests-integration/tests/traits_test.rs
@@ -16,7 +16,6 @@
 // under the License.
 
 use asyncband::barrier::Barrier;
-use asyncband::broadcast;
 use asyncband::condvar::Condvar;
 use asyncband::latch::Latch;
 use asyncband::mpsc;
@@ -82,10 +81,6 @@ fn public_types_are_send_and_sync() {
     assert_send_and_sync::<OwnedRwLockReadGuard<i64>>();
     assert_send_and_sync::<RwLockReadGuard<'_, i64>>();
     assert_send_and_sync::<RwLockWriteGuard<'_, i64>>();
-    assert_send_and_sync::<broadcast::overflow::Sender<i64>>();
-    assert_send_and_sync::<broadcast::overflow::Receiver<i64>>();
-    assert_send_and_sync::<broadcast::overflow::RecvError>();
-    assert_send_and_sync::<broadcast::overflow::TryRecvError>();
     assert_send_and_sync::<oneshot::SendError<i64>>();
     assert_send_and_sync::<oneshot::Sender<i64>>();
     assert_send_and_sync::<pool::bounded::Pool<PoolManager>>();
@@ -130,10 +125,6 @@ fn public_types_are_unpin() {
     assert_unpin::<RwLock<i64>>();
     assert_unpin::<RwLockReadGuard<'_, i64>>();
     assert_unpin::<RwLockWriteGuard<'_, i64>>();
-    assert_unpin::<broadcast::overflow::Sender<i64>>();
-    assert_unpin::<broadcast::overflow::Receiver<i64>>();
-    assert_unpin::<broadcast::overflow::RecvError>();
-    assert_unpin::<broadcast::overflow::TryRecvError>();
     assert_unpin::<oneshot::Sender<i64>>();
     assert_unpin::<oneshot::SendError<i64>>();
     assert_unpin::<oneshot::Receiver<i64>>();